Clean age column complete
added clean age and clean zip column, completed clean age column.

Note: I am currently opening the CSV in google sheets, then saving changes to a new CSV, then copying that CSV text into my local version of the orignal, then committing THAT.

If it sounds complicated, that's because it is, but if it's working, it works for me.